fix user create a Invoice

+++ src/main/java/com/example/springbootjdbc/Controllers/InvoiceController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.text.SimpleDateFormat;
+import java.util.Date;
 
 @Controller
 public class InvoiceController {
@@ -21,7 +22,7 @@ public class InvoiceController {
 
     @PostMapping("payment")
     public String createInvoice(@RequestParam String title, @RequestParam String description,
-                                @RequestParam String category, @RequestParam double price) {
+                                @RequestParam String category, @RequestParam int price) {
         new Invoice(title, description,
                 category, price);
         invoiceService.addInvoice(new Invoice(title, description, category, price));
